# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Korean Car Search API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Car {
  id: string;
  brand: string;
  model: string;
  year: number;
  /** Price in KRW millions */
  price: number;
  priceFormatted: string;
  /** Mileage in km */
  mileage: number;
  fuelType: string;
  transmission: string;
  bodyType: string;
  color: string;
  sunroof: boolean;
  imageUrl?: string;
  thumbnailUrl?: string;
  description?: string;
  features?: string[];
  /** Source website name */
  source: string;
  /** URL to the listing on the source website */
  sourceUrl: string;
  location?: string;
  createdAt?: string;
}

export interface CarSearchResponse {
  cars: Car[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface BrandsResponse {
  brands: string[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type SearchCarsParams = {
  /**
   * Search query (model name, brand)
   */
  query?: string;
  /**
   * Car brand (Hyundai, Kia, Genesis, SsangYong, etc.)
   */
  brand?: string;
  /**
   * Model year from
   */
  yearFrom?: number;
  /**
   * Model year to
   */
  yearTo?: number;
  /**
   * Has sunroof
   */
  sunroof?: boolean;
  /**
   * Transmission type (auto, manual)
   */
  transmission?: SearchCarsTransmission;
  /**
   * Fuel type
   */
  fuelType?: SearchCarsFuelType;
  /**
   * Body type
   */
  bodyType?: SearchCarsBodyType;
  /**
   * Minimum price in KRW (Korean Won millions)
   */
  priceMin?: number;
  /**
   * Maximum price in KRW (Korean Won millions)
   */
  priceMax?: number;
  /**
   * Maximum mileage in km
   */
  mileageMax?: number;
  /**
   * Car color
   */
  color?: string;
  /**
   * Page number
   */
  page?: number;
  /**
   * Results per page
   */
  limit?: number;
};

export type SearchCarsTransmission =
  (typeof SearchCarsTransmission)[keyof typeof SearchCarsTransmission];

export const SearchCarsTransmission = {
  auto: "auto",
  manual: "manual",
  any: "any",
} as const;

export type SearchCarsFuelType =
  (typeof SearchCarsFuelType)[keyof typeof SearchCarsFuelType];

export const SearchCarsFuelType = {
  gasoline: "gasoline",
  diesel: "diesel",
  hybrid: "hybrid",
  electric: "electric",
  any: "any",
} as const;

export type SearchCarsBodyType =
  (typeof SearchCarsBodyType)[keyof typeof SearchCarsBodyType];

export const SearchCarsBodyType = {
  sedan: "sedan",
  suv: "suv",
  hatchback: "hatchback",
  coupe: "coupe",
  van: "van",
  truck: "truck",
  any: "any",
} as const;