# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Korean Car Search API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: cars
    description: Car search operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /cars/search:
    get:
      operationId: searchCars
      tags: [cars]
      summary: Search Korean cars
      description: Search for cars with various filters
      parameters:
        - name: query
          in: query
          description: Search query (model name, brand)
          schema:
            type: string
        - name: brand
          in: query
          description: Car brand (Hyundai, Kia, Genesis, SsangYong, etc.)
          schema:
            type: string
        - name: yearFrom
          in: query
          description: Model year from
          schema:
            type: integer
        - name: yearTo
          in: query
          description: Model year to
          schema:
            type: integer
        - name: sunroof
          in: query
          description: Has sunroof
          schema:
            type: boolean
        - name: transmission
          in: query
          description: Transmission type (auto, manual)
          schema:
            type: string
            enum: [auto, manual, any]
        - name: fuelType
          in: query
          description: Fuel type
          schema:
            type: string
            enum: [gasoline, diesel, hybrid, electric, any]
        - name: bodyType
          in: query
          description: Body type
          schema:
            type: string
            enum: [sedan, suv, hatchback, coupe, van, truck, any]
        - name: priceMin
          in: query
          description: Minimum price in KRW (Korean Won millions)
          schema:
            type: integer
        - name: priceMax
          in: query
          description: Maximum price in KRW (Korean Won millions)
          schema:
            type: integer
        - name: mileageMax
          in: query
          description: Maximum mileage in km
          schema:
            type: integer
        - name: color
          in: query
          description: Car color
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Results per page
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarSearchResponse"
  /cars/brands:
    get:
      operationId: getCarBrands
      tags: [cars]
      summary: Get available car brands
      responses:
        "200":
          description: List of car brands
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandsResponse"
  /cars/{id}:
    get:
      operationId: getCarById
      tags: [cars]
      summary: Get car details by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Car details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "404":
          description: Car not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Car:
      type: object
      properties:
        id:
          type: string
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        price:
          type: integer
          description: Price in KRW millions
        priceFormatted:
          type: string
        mileage:
          type: integer
          description: Mileage in km
        fuelType:
          type: string
        transmission:
          type: string
        bodyType:
          type: string
        color:
          type: string
        sunroof:
          type: boolean
        imageUrl:
          type: string
        thumbnailUrl:
          type: string
        description:
          type: string
        features:
          type: array
          items:
            type: string
        source:
          type: string
          description: Source website name
        sourceUrl:
          type: string
          description: URL to the listing on the source website
        location:
          type: string
        createdAt:
          type: string
      required:
        - id
        - brand
        - model
        - year
        - price
        - priceFormatted
        - mileage
        - fuelType
        - transmission
        - bodyType
        - color
        - sunroof
        - source
        - sourceUrl
    CarSearchResponse:
      type: object
      properties:
        cars:
          type: array
          items:
            $ref: "#/components/schemas/Car"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - cars
        - total
        - page
        - limit
        - totalPages
    BrandsResponse:
      type: object
      properties:
        brands:
          type: array
          items:
            type: string
      required:
        - brands
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message