# lib/api-zod/src/generated/types/car.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Korean Car Search API
 * OpenAPI spec version: 0.1.0
 */

export interface Car {
  id: string;
  brand: string;
  model: string;
  year: number;
  /** Price in KRW millions */
  price: number;
  priceFormatted: string;
  /** Mileage in km */
  mileage: number;
  fuelType: string;
  transmission: string;
  bodyType: string;
  color: string;
  sunroof: boolean;
  imageUrl?: string;
  thumbnailUrl?: string;
  description?: string;
  features?: string[];
  /** Source website name */
  source: string;
  /** URL to the listing on the source website */
  sourceUrl: string;
  location?: string;
  createdAt?: string;
}