# lib/api-zod/src/generated/types/carSearchResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Korean Car Search API
 * OpenAPI spec version: 0.1.0
 */
import type { Car } from "./car";

export interface CarSearchResponse {
  cars: Car[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}